fixes: typos, hoisted imports that generate loops
