+++ model/coincident.py
@@ -7,7 +7,7 @@ from ..solver import Solver
 from ..global_data import WpReq
 from ..base.constants import SOLVER_GROUP_FIXED
 from .base_constraint import GenericConstraint
-from .utilities import slvs_entity_pointer, make_coincident
+from .utilities import slvs_entity_pointer
 from .categories import POINT, LINE
 from .workplane import SlvsWorkplane
 from .arc import SlvsArc
@@ -39,6 +39,7 @@ class SlvsCoincident(GenericConstraint, PropertyGroup):
         return WpReq.OPTIONAL
 
     def create_slvs_data(self, solvesys, group=SOLVER_GROUP_FIXED):
+        from ..utilities.data_handling import make_coincident
         return make_coincident(
             solvesys, self.entity1.py_data, self.entity2, self.get_workplane(), group
         )
