+++ model/sketch.py
@@ -6,7 +6,6 @@ from bpy.types import PropertyGroup
 from bpy.props import EnumProperty, BoolProperty, IntProperty, PointerProperty
 from bpy.utils import register_classes_factory
 
-from ..solver import solve_system
 from ..base.constants import SOLVER_GROUP_FIXED
 from .base_entity import SlvsGenericEntity
 from .utilities import slvs_entity_pointer
@@ -82,7 +81,7 @@ class SlvsSketch(SlvsGenericEntity, PropertyGroup):
     def draw_id(self, context):
         pass
 
-    def create_slvs_data(self, solvesys, group=SSOLVER_GROUP_FIXED):
+    def create_slvs_data(self, solvesys, group=SOLVER_GROUP_FIXED):
         pass
 
     def remove_objects(self):
@@ -100,6 +99,7 @@ class SlvsSketch(SlvsGenericEntity, PropertyGroup):
         return bpyEnum(solver_state_items, identifier=self.solver_state)
 
     def solve(self, context):
+        from ..solver import solve_system
         return solve_system(context, sketch=self)
 
     @classmethod
