+++ model/tangent.py
@@ -6,7 +6,7 @@ from bpy.utils import register_classes_factory
 from ..base.constants import SOLVER_GROUP_FIXED
 from ..global_data import WpReq
 from .base_constraint import GenericConstraint
-from .utilities import slvs_entity_pointer, make_coincident, get_connection_point
+from .utilities import slvs_entity_pointer, get_connection_point
 from .categories import CURVE
 from .line_2d import SlvsLine2D
 from .arc import SlvsArc
@@ -50,6 +50,9 @@ class SlvsTangent(GenericConstraint, PropertyGroup):
                 )
 
         elif isinstance(e2, SlvsLine2D):
+            # avoid circular dependency
+            from ..utilities.data_handling import make_coincident
+
             orig = e2.p1.co
             coords = (e1.ct.co - orig).project(e2.p2.co - orig) + orig
             params = [solvesys.addParamV(v, group) for v in coords]
